better consistency of time units
timestamps in milliseconds use qint64 type; in seconds,
use qreal, because fractional seconds are possible.
sampleNearest(time) uses the same form of time as the samples
themselves: seconds since startup.  Since that's what we intend,
get the startup time once and store it statically; that way
time is consistent across any number of graphs.

+++ linegraph.h
@@ -56,7 +56,7 @@ public:
     void setWireframe(bool wireframe);
 
     Q_INVOKABLE QJSValue sampleNearestX(qreal x);
-    Q_INVOKABLE qreal xAtTime(qint64 time);
+    Q_INVOKABLE qreal xAtTime(qreal time);
 
     static void registerMetaType();
